core: refactor - (models, utilities)
