+++ backend/src/api/bot/mod.rs
@@ -3,3 +3,4 @@ pub mod models;
 pub mod photo;
 pub mod preprocessing;
 pub mod redis;
+pub mod utilities;
