+++ backend/src/api/microservices/meilisearch.rs
@@ -1,11 +1,10 @@
 use super::database::{
-    core::convert_db_items,
     init::DatabaseQueries,
-    schema::{columns::items, tables},
+    schema::{columns::boiler_swap::items, tables},
 };
 use crate::{
     AppError,
-    api::models::ItemRow,
+    api::{models::ItemRow, web::swap::convert_db_items},
     config::{read_secret, try_load},
 };
 use meilisearch_sdk::{
@@ -47,7 +46,7 @@ pub async fn init_meilisearch(
     let settings = init_settings();
 
     meili_client
-        .index(tables::ITEMS)
+        .index(tables::boiler_swap::ITEMS)
         .set_settings(&settings)
         .await
         .unwrap();
@@ -60,7 +59,7 @@ pub async fn init_meilisearch(
             session_clone,
             queries_clone,
             client_clone,
-            tables::ITEMS,
+            tables::boiler_swap::ITEMS,
             items::ITEM_ID,
             item_counter_clone,
         )
@@ -84,7 +83,7 @@ pub async fn reindex(
 
     loop {
         let (query_result, paging_state_response) = database_session
-            .execute_single_page(&database_queries.get_items, &[], paging_state)
+            .execute_single_page(&database_queries.boiler_swap.get_items, &[], paging_state)
             .await?;
 
         let row_result = query_result.into_rows_result()?;
@@ -127,6 +126,7 @@ where
         .await?
         .wait_for_completion(&meili_client, None, None)
         .await?;
+
     Ok(())
 }
 
@@ -141,6 +141,7 @@ pub async fn delete_item(
         .await?
         .wait_for_completion(&meili_client, None, None)
         .await?;
+
     Ok(())
 }
 
@@ -151,6 +152,7 @@ pub async fn clear_index(meili_client: Arc<Client>, index_name: &str) -> anyhow:
         .await?
         .wait_for_completion(&meili_client, None, None)
         .await?;
+
     Ok(())
 }
